processing-service/requirements.txt
fastapi
uvicorn
requests
transformers
sentence_transformers
spacy
psycopg2-binary
python-dotenv
nltk